requirements_win64.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
attrs==23.2.0
audioread==3.0.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.4.2
datasets==2.19.1
decorator==5.1.1
dill==0.3.8
filelock==3.14.0
frozenlist==1.4.1
fsspec==2024.3.1
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.4
googleapis-common-protos==1.63.0
grpcio==1.64.0
grpcio-status==1.62.2
gTTS==2.5.1
httplib2==0.22.0
huggingface-hub==0.23.2
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.42.0
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
numba==0.59.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
platformdirs==4.2.2
playsound==1.3.0
pooch==1.8.1
proto-plus==1.23.0
protobuf==4.25.3
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
pyparsing==3.1.2
pypiwin32==223
python-dateutil==2.9.0.post0
pyttsx3==2.90
pytz==2024.1
pywin32==306
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.2
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.13.1
sentencepiece==0.2.0
six==1.16.0
sounddevice==0.4.7
soundfile==0.12.1
soxr==0.3.7
SpeechRecognition==3.10.4
sympy==1.12
tbb==2021.12.0
threadpoolctl==3.5.0
tokenizers==0.19.1
torch==2.3.0
tqdm==4.66.4
transformers==4.41.1
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
xxhash==3.4.1
yarl==1.9.4
